requirements.txt
flask
redis
qrcode[pil]
python-dotenv
gunicorn